Experience Updates, Drop X Config, Double EXP to Int EXP

+++ server/plugins/com/openrsc/server/plugins/skills/LootItem.java
@@ -2,17 +2,17 @@ package com.openrsc.server.plugins.skills;
 
 
 public class LootItem implements Comparable<LootItem> {
-	private double chance;
+	private int chance;
 	private final int id;
 	private final int amount;
 
-	public LootItem(int id, int amount, double chance) {
+	public LootItem(int id, int amount, int chance) {
 		this.id = id;
 		this.amount = amount;
 		this.chance = chance;
 	}
 
-	public double getChance() {
+	public int getChance() {
 		return chance;
 	}
